yandexbot/requirements.txt
alembic==1.10.4
anyio==3.6.2
certifi==2022.12.7
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
python-telegram-bot==20.2
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.11
typing_extensions==4.5.0